use size_t as cu_num type

+++ src/include/miopen/conv/asm_implicit_gemm.hpp
@@ -130,8 +130,8 @@ static inline size_t ComputeGemmKGlobalSplitsWith2DMerge(size_t current_grid_siz
                                                          // size_t gemm_k_per_block,
                                                          size_t occupancy)
 {
-    int num_cu               = 120;
-    int gemm_k_global_splits = num_cu * occupancy / current_grid_size;
+    size_t num_cu               = 120;
+    size_t gemm_k_global_splits = num_cu * occupancy / current_grid_size;
     // int gemm_k_per_wg = math::integer_divide_ceil(merge_dimension / gemm_k_global_splits);
     // gemm_k_per_wg = (gemm_k_per_wg + gemm_k_per_block - 1) / gemm_k_per_block * gemm_k_per_block;
     // gemm_k_global_splits = math::integer_divide_ceil(merge_dimension / gemm_k_per_wg);
